[chore] demoSection 생성, popup coordinator 추가 #50
