+++ FindTown/FindTown/Presentation/MyPageScene/Flows/MyPageCoordinator.swift
@@ -56,4 +56,14 @@ extension MyPageCoordinator: MyPageViewModelDelegate {
         guard let navigationController = navigationController else { return }
         navigationController.pushViewController(myTownReviewScene(), animated: true)
     }
+    
+    func popUpSignout() {
+        guard let navigationController = navigationController else { return }
+        navigationController.showAlertSuccessCancelPopUp(title: "동네한입", message: "로그아웃 하시겠어요?", successButtonText: "로그아웃", cancelButtonText: "취소", successButtonAction: { print("로그아웃") })
+    }
+    
+    func popUpWithDraw() {
+        guard let navigationController = navigationController else { return }
+        navigationController.showAlertSuccessCancelPopUp(title: "회원탈퇴", message: "탈퇴 시 모든 정보는 삭제됩니다.\n(단, 동네후기는 제외)\n정말로 탈퇴하시겠어요?", successButtonText: "탈퇴", cancelButtonText: "취소", successButtonAction: { print("탈퇴" )})
+    }
 }
