+++ FindTown/FindTown/Presentation/MyPageScene/MyPage/MyPageViewModel.swift
@@ -14,11 +14,15 @@ import RxRelay
 protocol MyPageViewModelDelegate {
     func goToChangeNickname()
     func goToMyTownReview()
+    func popUpSignout()
+    func popUpWithDraw()
 }
 
 protocol MyPageViewModelType {
     func goToChangeNickname()
     func goToMyTownReview()
+    func popUpSignout()
+    func popUpWithDraw()
 }
 
 final class MyPageViewModel: BaseViewModel {
@@ -40,7 +44,7 @@ final class MyPageViewModel: BaseViewModel {
         delegate: MyPageViewModelDelegate
     ) {
         self.delegate = delegate
-    
+        
         super.init()
         self.bind()
     }
@@ -68,65 +72,12 @@ extension MyPageViewModel: MyPageViewModelType {
     func goToMyTownReview() {
         delegate.goToMyTownReview()
     }
-}
-
-enum MyPageDemoData {
-    static let dataSource = [
-        MyPageSection.topHeader([ ]),
-        MyPageSection.support(
-            [
-                .init(image: UIImage(),
-                      title: "지원",
-                      action: { print("") }),
-                .init(image: UIImage(named: "contact") ?? UIImage(),
-                      title: "문의하기",
-                      action: { print("문의하기") }),
-                .init(image: UIImage(named: "suggestion") ?? UIImage(),
-                      title: "제안하기",
-                      action: { print("제안하기") }),
-            ]
-        ),
-        MyPageSection.info(
-            [
-                .init(title: "정보",
-                      index: 0,
-                      action: { print("정보") }),
-                .init(title: "이용약관",
-                      index: 1,
-                      action: { print("이용약관") }),
-                .init(title: "개인정보처리 방침",
-                      index: 2,
-                      action: { print("개인정보처리 방침") }),
-                .init(title: "앱 버전",
-                      index: 3,
-                      action: { print("앱 버전") }),
-                .init(title: "로그아웃",
-                      index: 4,
-                      action: { print("로그아웃") }),
-                .init(title: "회원탈퇴",
-                      index: 5,
-                      action: { print("회원탈퇴") }),
-            ]
-        ),
-    ]
-}
-
-enum MyPageSection {
-    struct TopHeader { }
     
-    struct Support {
-        let image: UIImage
-        let title: String
-        let action: (() -> ())
+    func popUpSignout() {
+        delegate.popUpSignout()
     }
     
-    struct Info {
-        let title: String
-        let index: Int
-        let action: (() -> ())
+    func popUpWithDraw() {
+        delegate.popUpWithDraw()
     }
-    
-    case topHeader([TopHeader])
-    case support([Support])
-    case info([Info])
 }
